+++ lib/canon.dart
@@ -1,3 +1,4 @@
+export 'package:canon_codec/canon_codec.dart' hide ListCodec;
 export 'src/nav_graph.dart' hide Nav;
 export 'src/screen_node.dart'
     hide GrammarNode, NavSpec, NavResolution, resolveGo, resolvePop;